requirements.txt
djangorestframework==3.11.0
Django==2.2.8
psycopg2-binary==2.8.4
packaging==20.1
django-cors-headers==3.2.1